feat(basic-emulator): add support for mac
* use OpenGL 4.1 instead of 4.5
* define darwin architecture

+++ source/arch/thread.h
@@ -12,7 +12,7 @@ typedef unsigned long (*ThreadRunner)(void *);
 
 #else
 
-typedef u32 Thread;
+typedef usize Thread;
 typedef void *(*ThreadRunner)(void *);
 
 #endif
